Fix package name for Azure tree nodes
