+++ PluginsAndFeatures/azure-toolkit-for-eclipse/com.microsoft.azuretools.azureexplorer/src/com/microsoft/azure/toolkit/eclipse/explorer/AzureExplorer.java
@@ -3,7 +3,7 @@
  * Licensed under the MIT License. See License.txt in the project root for license information.
  */
 
-package com.microsoft.azure.toolkit.intellij.explorer;
+package com.microsoft.azure.toolkit.eclipse.explorer;
 
 import java.util.Arrays;
